+++ .gitignore
@@ -36,4 +36,4 @@ out/
 ### VS Code ###
 .vscode/
 
-/src/main/resources/application-oauth.yml
+/src/main/resources/application-aws.yml
