--- src/main/java/com/example/webtoon/User/domain/model/BaseEntity.java
@@ -1,20 +1,0 @@
-package com.example.webtoon.User.domain.model;
-
-import java.time.LocalDateTime;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(value = {AuditingEntityListener.class})
-public class BaseEntity {
-
-    @CreatedDate
-    private LocalDateTime createdAt;
-    @LastModifiedDate
-    private LocalDateTime modifiedAt;
-}
