--- src/main/java/com/example/webtoon/User/security/JwtAuthenticationFilter.java
@@ -1,53 +1,0 @@
-package com.example.webtoon.User.security;
-
-
-import java.io.IOException;
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.util.ObjectUtils;
-import org.springframework.util.StringUtils;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-
-    public static final String TOKEN_HEADER = "Authorization";
-    public static final String TOKEN_PREFIX = "Bearer ";
-
-    private final TokenProvider tokenProvider;
-
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
-        FilterChain filterChain) throws ServletException, IOException {
-        String token = this.resolveTokenFromRequest(request);
-
-        if (StringUtils.hasText(token) && tokenProvider.validateToken(token)) {
-            Authentication auth = tokenProvider.getAuthentication(token);
-            SecurityContextHolder.getContext().setAuthentication(auth);
-
-            log.info("auth : " + auth.getAuthorities());
-
-            log.info("[{}] -> {}", this.tokenProvider.getUserId(token), request.getRequestURI());
-        }
-        filterChain.doFilter(request, response);
-    }
-
-    private String resolveTokenFromRequest(HttpServletRequest request) {
-        String token = request.getHeader(TOKEN_HEADER);
-
-        if ((ObjectUtils.isEmpty(token) || token.startsWith(TOKEN_PREFIX)) && token != null) {
-            return token.substring(TOKEN_PREFIX.length());
-        }
-        return "";
-    }
-}
