--- src/main/java/com/example/webtoon/User/type/Authority.java
@@ -1,6 +1,0 @@
-package com.example.webtoon.User.type;
-
-public enum Authority {
-    ROLE_USER,
-    ROLE_MANAGER
-}
