--- src/main/java/com/example/webtoon/exception/ExceptionController.java
@@ -1,33 +1,0 @@
-package com.example.webtoon.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.ToString;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-@Slf4j
-public class ExceptionController {
-
-    @ExceptionHandler({
-        GlobalException.class
-    })
-    public ResponseEntity<ExceptionResponse> customRequestException(final GlobalException c) {
-        log.warn("api Exception : {}", c.getErrorCode());
-        return ResponseEntity.badRequest()
-            .body(new ExceptionResponse(c.getMessage(), c.getErrorCode()));
-    }
-
-
-    @Getter
-    @ToString
-    @AllArgsConstructor
-    public static class ExceptionResponse {
-
-        private String message;
-        private ErrorCode errorCode;
-    }
-}
