--- src/main/java/com/example/webtoon/exception/GlobalException.java
@@ -1,14 +1,0 @@
-package com.example.webtoon.exception;
-
-import lombok.Getter;
-
-@Getter
-public class GlobalException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public GlobalException(ErrorCode errorCode) {
-        super(errorCode.getDetail());
-        this.errorCode = errorCode;
-    }
-}
