+++ src/main/java/com/example/webtoon/global/exception/GlobalException.java
@@ -1,0 +1,14 @@
+package com.example.webtoon.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class GlobalException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+
+    public GlobalException(ErrorCode errorCode) {
+        super(errorCode.getDetail());
+        this.errorCode = errorCode;
+    }
+}
