+++ src/main/java/com/example/webtoon/user/domain/PrincipalDetails.java
@@ -1,0 +1,55 @@
+package com.example.webtoon.user.domain;
+
+import com.example.webtoon.user.domain.model.User;
+import java.util.ArrayList;
+import java.util.Collection;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+public class PrincipalDetails implements UserDetails {
+
+    private User user;
+
+    public PrincipalDetails(User user) {
+        this.user = user;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        SimpleGrantedAuthority userAuthority = new SimpleGrantedAuthority(user.getRole().name());
+        Collection<GrantedAuthority> authorities = new ArrayList<>();
+        authorities.add(userAuthority);
+        return authorities;
+    }
+
+    @Override
+    public String getPassword() {
+        return user.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getEmail();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return !isEnabled();
+    }
+}
