+++ src/main/java/com/example/webtoon/user/domain/dto/SignInForm.java
@@ -1,0 +1,12 @@
+package com.example.webtoon.user.domain.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class SignInForm {
+
+    private String email;
+    private String password;
+}
