+++ src/main/java/com/example/webtoon/user/domain/dto/TokenResponse.java
@@ -1,0 +1,10 @@
+package com.example.webtoon.user.domain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class TokenResponse {
+    private String token;
+}
