+++ src/main/java/com/example/webtoon/user/domain/repository/UserRepository.java
@@ -1,0 +1,14 @@
+package com.example.webtoon.user.domain.repository;
+
+import com.example.webtoon.user.domain.model.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findUserByEmail(String email);
+
+    boolean existsByEmail(String email);
+}
