+++ src/main/java/com/example/webtoon/user/service/PrincipalService.java
@@ -1,0 +1,26 @@
+package com.example.webtoon.user.service;
+
+import com.example.webtoon.user.domain.PrincipalDetails;
+import com.example.webtoon.user.domain.model.User;
+import com.example.webtoon.user.domain.repository.UserRepository;
+import com.example.webtoon.global.exception.ErrorCode;
+import com.example.webtoon.global.exception.GlobalException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class PrincipalService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String userEmail) throws UsernameNotFoundException {
+        User user = userRepository.findUserByEmail(userEmail)
+            .orElseThrow(() -> new GlobalException(ErrorCode.NOT_FOUND_USER));
+        return new PrincipalDetails(user);
+    }
+}
