+++ src/main/java/com/example/webtoon/user/type/Authority.java
@@ -1,0 +1,6 @@
+package com.example.webtoon.user.type;
+
+public enum Authority {
+    ROLE_USER,
+    ROLE_AUTHOR
+}
