+++ src/main/java/com/example/webtoon/webtoon/domain/model/WebtoonChapter.java
@@ -1,0 +1,37 @@
+package com.example.webtoon.webtoon.domain.model;
+
+import com.example.webtoon.global.domain.BaseEntity;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class WebtoonChapter extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @ManyToOne
+    private Webtoon webtoon;
+    private String chapterName;
+    private Integer chapterCount;
+    @Builder.Default
+    @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
+    @JoinColumn(name = "WEBTOONCHAPTER_ID")
+    private List<WebtoonImg> webtoonImgs = new ArrayList<>();
+}
