+++ src/main/java/com/example/webtoon/webtoon/domain/model/WebtoonHashtag.java
@@ -1,0 +1,46 @@
+package com.example.webtoon.webtoon.domain.model;
+
+import java.util.Objects;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class WebtoonHashtag {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+
+    private WebtoonHashtag(String hashtagName){
+        this.name = hashtagName;
+    }
+
+    public String getName(){
+        return name;
+    }
+
+    public static WebtoonHashtag of(String hashtagName){
+        return new WebtoonHashtag(hashtagName);
+    }
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        WebtoonHashtag that = (WebtoonHashtag) o;
+        return name.equals(that.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name);
+    }
+}
