+++ src/main/java/com/example/webtoon/webtoon/domain/tyoe/Day.java
@@ -1,0 +1,11 @@
+package com.example.webtoon.webtoon.domain.tyoe;
+
+public enum Day {
+    MON,
+    TUE,
+    WED,
+    THU,
+    FRI,
+    SAT,
+    SUN
+}
